+++ test/controllers/custom_registrations_controller_test.rb
@@ -3,7 +3,7 @@ require 'test_helper'
 class CustomRegistrationsControllerTest < Devise::ControllerTestCase
   tests Custom::RegistrationsController
 
-  include Devise::TestHelpers
+  include Devise::Test::ControllerHelpers
 
   setup do
     request.env["devise.mapping"] = Devise.mappings[:user]
