+++ test/controllers/custom_strategy_test.rb
@@ -27,7 +27,7 @@ end
 class CustomStrategyTest < Devise::ControllerTestCase
   tests CustomStrategyController
 
-  include Devise::TestHelpers
+  include Devise::Test::ControllerHelpers
 
   setup do
     Warden::Strategies.add(:custom_strategy, CustomStrategy)
