+++ test/controllers/passwords_controller_test.rb
@@ -2,7 +2,7 @@ require 'test_helper'
 
 class PasswordsControllerTest < Devise::ControllerTestCase
   tests Devise::PasswordsController
-  include Devise::TestHelpers
+  include Devise::Test::ControllerHelpers
 
   setup do
     request.env["devise.mapping"] = Devise.mappings[:user]
